+++ liveness-server-typescript/src/liveness/common.ts
@@ -1,6 +1,11 @@
-
 export enum LivenessMode {
     DetectLiveness = 'detectLiveness',
     DetectLivenessWithVerify = 'detectLivenessWithVerify'
   }
   
+export type LivenessResult = {
+  sessionId: string;
+  status: string;
+  livenessDecision?: string;
+  verifyMatchDecision?: boolean;
+};
